add botic7pre1

+++ build_kernel.sh
@@ -88,9 +88,9 @@ config_comparsion () {
 
 copy_defconfig () {
 	cd "${DIR}/KERNEL" || exit
+	make ARCH=arm CROSS_COMPILE="${CC}" distclean
 	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" "${config}"
-
-	cp -v "${DIR}/patches/defconfig" .config
+	cp -v .config "${DIR}/patches/defconfig"
 	cd "${DIR}/" || exit
 }
 
@@ -130,6 +130,12 @@ make_kernel () {
 	fi
 
 	KERNEL_UTS=$(cat "${DIR}/KERNEL/include/generated/utsrelease.h" | awk '{print $3}' | sed 's/\"//g' )
+
+	if [ "${FAST_REBUILD}" ]; then
+		echo "${KERNEL_UTS}" > kernel_version
+		echo "Fast rebuild DONE."
+		exit
+	fi
 
 	if [ -f "${DIR}/deploy/${KERNEL_UTS}.${image}" ] ; then
 		rm -rf "${DIR}/deploy/${KERNEL_UTS}.${image}" || true
@@ -215,7 +221,10 @@ make_dtbs_pkg () {
 	make_pkg
 }
 
-/bin/sh -e "${DIR}/tools/host_det.sh" || { exit 1 ; }
+FAST_REBUILD=$1
+if [ ! "${FAST_REBUILD}" ]; then
+	/bin/sh -e "${DIR}/tools/host_det.sh" || { exit 1 ; }
+fi
 
 if [ ! -f "${DIR}/system.sh" ] ; then
 	cp -v "${DIR}/system.sh.sample" "${DIR}/system.sh"
@@ -257,7 +266,18 @@ fi
 export LINUX_GIT
 
 #unset FULL_REBUILD
-FULL_REBUILD=1
+if [ "${FULL_REBUILD}" -o ! "${FAST_REBUILD}" ]; then
+	echo "Full rebuild will start in 10 seconds..."
+	sleep 10
+	FULL_REBUILD=1
+elif [ "${FAST_REBUILD}" = 0 ]; then
+	FULL_REBUILD=
+	AUTO_BUILD=
+else
+	FULL_REBUILD=
+	AUTO_BUILD=1
+fi
+
 if [ "${FULL_REBUILD}" ] ; then
 	/bin/sh -e "${DIR}/scripts/git.sh" || { exit 1 ; }
 	cp -v "${DIR}/KERNEL/scripts/package/builddeb" "${DIR}/3rdparty/packaging/"
@@ -269,7 +289,8 @@ if [ "${FULL_REBUILD}" ] ; then
 
 	patch_kernel
 	if [ ! "${AUTO_BUILD}" ] ; then
-		config_comparsion
+		## No config comparsion for Botic
+		false && config_comparsion
 	fi
 	copy_defconfig
 fi
